🎨 응답값 수정

+++ app/domains/company_users/service.py
@@ -2,10 +2,15 @@ from fastapi import HTTPException, status
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.domains.company_users.schemas import (CompanyUserRequest,
-                                               CompanyUserUpdateRequest)
-from app.domains.company_users.utiles import (check_business_number_valid,
-                                              hash_password, verify_password)
+from app.domains.company_users.schemas import (
+    CompanyUserRequest,
+    CompanyUserUpdateRequest,
+)
+from app.domains.company_users.utiles import (
+    check_business_number_valid,
+    hash_password,
+    verify_password,
+)
 from app.domains.users.service import create_access_token, create_refresh_token
 from app.models import CompanyInfo, CompanyUser
 
@@ -94,7 +99,7 @@ async def register_company_user(db: AsyncSession, payload: CompanyUserRequest):
         # 기업 유저(담당자) 정보 저장
         company_user = await create_company_user(db, payload, company_info.id)
 
-        return company_user
+        return success_response("회원가입이 완료되었습니다.", company_user)
 
     except Exception as e:
         await db.rollback()
@@ -119,14 +124,19 @@ async def login_company_user(db: AsyncSession, email: str, password: str):
             status_code=status.HTTP_401_UNAUTHORIZED,
             detail="비밀번호가 올바르지 않습니다.",
         )
-    access_token = await create_access_token(data={"sub": company_user.id})
-    refresh_token = await create_refresh_token(data={"sub": company_user.id})
+    access_token = await create_access_token(data={"sub": company_user.email})
+    refresh_token = await create_refresh_token(data={"sub": company_user.email})
 
-    return {
-        "company_user": company_user,
+    return success_response("로그인이 완료되었습니다.", {
         "access_token": access_token,
         "refresh_token": refresh_token,
-    }
+        "token_type": "bearer",
+        "company_user": {
+            "company_user_id": company_user.id,
+            "email": company_user.email,
+            "cem_name": company_user.manager_name,
+        }
+    })
 
 
 # 기업 회원 정보 조회(마이페이지)
@@ -142,7 +152,7 @@ async def get_company_user_mypage(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="해당 기업 정보를 찾을 수 없습니다.",
         )
-    return {
+    return success_response("기업 회원 정보 조회 성공",{
         "company_info": {
             "company_user": current_user.company.company_name,  # 기업이름
             "manager_name": current_user.manager_name,  # 담당자이름
@@ -160,7 +170,7 @@ async def get_company_user_mypage(
             }
             for job_posting in current_user.job_postings
         ],
-    }
+    })
 
 
 # 기업 회원 정보수정
@@ -190,28 +200,23 @@ async def update_company_user(
 
     await db.commit()
     await db.refresh(current_user)
-    return current_user
+    return success_response("기업 정보가 수정되었습니다.",current_user)
+
 
 # 기업 회원 탈퇴
-async def delete_company_user(db: AsyncSession, company_user_id: int,current_user: CompanyUser):
+async def delete_company_user(
+    db: AsyncSession, company_user_id: int, current_user: CompanyUser
+):
 
     if current_user.id != company_user_id:
         raise HTTPException(
-            status_code=status.HTTP_403_FORBIDDEN,
-            detail="권한이 없습니다."
+            status_code=status.HTTP_403_FORBIDDEN, detail="권한이 없습니다."
         )
     user = await db.get(CompanyUser, company_user_id)
     if not user:
         raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="회원 정보가 없습니다."
+            status_code=status.HTTP_404_NOT_FOUND, detail="회원 정보가 없습니다."
         )
     if company_user_id:
-        await db.delete(
-            select(CompanyUser).filter_by(id=company_user_id)
-        )
-    return {
-        "status": "success",
-        "message": "회원 탈퇴가 정상적으로 처리 되었습니다."
-    }
-
+        await db.delete(select(CompanyUser).filter_by(id=company_user_id))
+    return {"status": "success", "message": "회원 탈퇴가 정상적으로 처리 되었습니다."}
